requirements.txt
luma.oled
Pillow
smbus2
psutil